chore: move prettier format to utils
