+++ scripts/apidoc/utils.ts
@@ -1,5 +1,8 @@
 import { resolve } from 'node:path';
+import type { Options } from 'prettier';
+import { format } from 'prettier';
 import * as TypeDoc from 'typedoc';
+import prettierConfig from '../../.prettierrc.cjs';
 import {
   DefaultParameterAwareSerializer,
   parameterDefaultReader,
@@ -44,3 +47,31 @@ export function newTypeDocApp(): TypeDoc.Application {
 export function patchProject(project: TypeDoc.ProjectReflection): void {
   patchProjectParameterDefaults(project);
 }
+
+/**
+ * Formats markdown contents.
+ *
+ * @param text The text to format.
+ */
+export function formatMarkdown(text: string): string {
+  return format(text, prettierMarkdown);
+}
+
+/**
+ * Formats typedoc contents.
+ *
+ * @param text The text to format.
+ */
+export function formatTypescript(text: string): string {
+  return format(text, prettierTypescript);
+}
+
+const prettierMarkdown: Options = {
+  ...prettierConfig,
+  parser: 'markdown',
+};
+
+const prettierTypescript: Options = {
+  ...prettierConfig,
+  parser: 'typescript',
+};
